Add JWT authentication and task ownership enforcement, update README

+++ routers/tasks.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session
 from database import SessionLocal 
 import models
 import schemas 
+from security import get_current_user
 
 router = APIRouter(
     prefix = "/tasks",
@@ -19,12 +20,12 @@ def get_db():
 
 # create task
 @router.post("/", response_model=schemas.TaskPublic, status_code = status.HTTP_201_CREATED) 
-def create_task(task: schemas.TaskCreate, db: Session = Depends(get_db)):
+def create_task(task: schemas.TaskCreate, db: Session = Depends(get_db), current_user: schemas.UserPublic = Depends(get_current_user)): 
     db_task = models.Task(
         title=task.title,
         description=task.description,
         completed=False,
-        user_id=task.user_id
+        user_id=current_user.id
     )
     db.add(db_task)
     db.commit()
@@ -33,8 +34,9 @@ def create_task(task: schemas.TaskCreate, db: Session = Depends(get_db)):
 
 # read all tasks
 @router.get("/", response_model=list[schemas.TaskPublic])
-def get_tasks(db: Session = Depends(get_db)):
-    return db.query(models.Task).all()
+def get_tasks(db: Session = Depends(get_db), current_user: schemas.UserPublic = Depends(get_current_user)): 
+    tasks = db.query(models.Task).filter(models.Task.user_id == current_user.id).all()
+    return tasks 
 
 # read task by id
 @router.get("/{task_id}", response_model=schemas.TaskPublic)
@@ -46,10 +48,12 @@ def get_task(task_id: int, db: Session = Depends(get_db)):
 
 # update task
 @router.put("/{task_id}", response_model=schemas.TaskPublic)
-def update_task(task_id: int, updated_task: schemas.TaskCreate, db: Session = Depends(get_db)):
+def update_task(task_id: int, updated_task: schemas.TaskCreate, db: Session = Depends(get_db), current_user: schemas.UserPublic = Depends(get_current_user)):
     task = db.query(models.Task).filter(models.Task.id == task_id).first()
     if not task:
-        raise HTTPException(status_code=404, detail="Task not found")
+        raise HTTPException(status_code=404, detail="Task not found") 
+    if task.user_id != current_user.id: 
+        raise HTTPException(status_code=403, detail="Not authorized to update this task!") 
     task.title = updated_task.title
     task.description = updated_task.description
     task.user_id = updated_task.user_id
@@ -59,9 +63,11 @@ def update_task(task_id: int, updated_task: schemas.TaskCreate, db: Session = De
 
 # delete task
 @router.delete("/{task_id}", status_code = status.HTTP_204_NO_CONTENT)
-def delete_task(task_id: int, db: Session = Depends(get_db)):
+def delete_task(task_id: int, db: Session = Depends(get_db), current_user: schemas.UserPublic = Depends(get_current_user)):
     task = db.query(models.Task).filter(models.Task.id == task_id).first()
     if not task:
         raise HTTPException(status_code=404, detail="Task not found")
+    if  task.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to delete this task!")
     db.delete(task)
     db.commit()
